+++ src/RcsbFv/RcsbFvRow/RcsbFvRowTitle.tsx
@@ -2,30 +2,35 @@ import * as React from "react";
 import {RcsbFvDefaultConfigValues} from "../RcsbFvConfig/RcsbFvDefaultConfigValues";
 import * as classes from "../RcsbFvStyles/RcsbFvRow.module.scss";
 import {RcsbFvRowConfigInterface} from "../RcsbFvConfig/RcsbFvConfigInterface";
-import {ReactElement} from "react";
+import {MouseEvent} from "react";
 
 /**Board track title cell React component interface*/
 interface RcsbFvRowTitleInterface {
-    data: RcsbFvRowConfigInterface;
-    rowTitleHeight: number;
+    readonly data: RcsbFvRowConfigInterface;
+    readonly rowTitleHeight: number;
 }
 
-export class RcsbFvRowTitle extends React.Component <RcsbFvRowTitleInterface, {}> {
+interface RcsbFvRowTitleInterState {
+    readonly expandTitle: boolean;
+}
+
+export class RcsbFvRowTitle extends React.Component <RcsbFvRowTitleInterface, RcsbFvRowTitleInterState> {
 
     configData : RcsbFvRowConfigInterface;
+    readonly state = {
+        expandTitle: false
+    };
     private readonly PADDING_RIGHT: number = 5;
 
     constructor(props: RcsbFvRowTitleInterface) {
         super(props);
         this.configData = this.props.data;
-        this.state = {
-
-        };
     }
 
     render(){
         const height: number = (this.configStyle().height as number);
         const trackTitle: string = typeof this.configData?.rowTitle === "string" ? this.configData.rowTitle : (typeof this.configData?.rowTitle === "object" ? this.configData.rowTitle.visibleTex : "");
+        const rowMark: JSX.Element = (<div style={{display:"inline-block", width:6, height:6, marginBottom: 4, marginRight:5}} className={classes.rcsbFvRowMark} ><div/></div>);
         if(typeof this.configData.rowPrefix === "string" && this.configData.rowPrefix.length > 0 && this.configData.fitTitleWidth){
             const prefixWidth: number = Math.round(((this.configData.rowPrefix.length/this.configData.rowPrefix.concat(trackTitle).length)*(this.configStyle().width as number)));
             const titleWidth: number = (this.configStyle().width as number)-prefixWidth;
@@ -36,8 +41,15 @@ export class RcsbFvRowTitle extends React.Component <RcsbFvRowTitleInterface, {}
                         this.setTitle() != null ? <div style={this.configTitleFlagColorStyle()}
                                                        className={classes.rcsbFvRowTitleProvenanceFlag}/> : null
                     }
-                    <div style={{...style, float:"right", display:"inline-block"}}><div className={classes.rcsbFvRowTitleText} style={{lineHeight:height+"px"}}>{this.setTitle()}</div></div>
-                    <div style={{height:height, float:"right", display:"inline-block"}}><div className={classes.rcsbFvRowTitleText}  style={{lineHeight:height+"px"}}>{this.configData.rowPrefix}</div></div>
+                    <div style={{...style, float:"right", display:"inline-block"}}>
+                        <div className={classes.rcsbFvRowTitleText} style={{lineHeight:height+"px"}}>{this.setTitle()}</div>
+                    </div>
+                    <div style={{height:height, float:"right", display:"inline-block"}}>
+                        <div className={classes.rcsbFvRowTitleText}  style={{lineHeight:height+"px", display:"inline-block"}}>
+                            {rowMark}
+                            {this.configData.rowPrefix}
+                        </div>
+                    </div>
                 </div>
             );
         }else if(typeof this.configData.rowPrefix === "string" && this.configData.rowPrefix.length > 0){
@@ -47,7 +59,10 @@ export class RcsbFvRowTitle extends React.Component <RcsbFvRowTitleInterface, {}
                         this.setTitle() != null ? <div style={this.configTitleFlagColorStyle()}
                                                        className={classes.rcsbFvRowTitleProvenanceFlag}/> : null
                     }
-                    <div className={classes.rcsbFvRowTitleText} style={{lineHeight:height+"px", paddingRight:this.PADDING_RIGHT}}>{this.configData.rowPrefix+" "}{this.setTitle()}</div>
+                    <div className={classes.rcsbFvRowTitleText} style={{lineHeight:height+"px", paddingRight:this.PADDING_RIGHT}}>
+                        {rowMark}
+                        {this.configData.rowPrefix+" "}{this.setTitle()}
+                    </div>
                 </div>
             );
         }else {
@@ -57,7 +72,10 @@ export class RcsbFvRowTitle extends React.Component <RcsbFvRowTitleInterface, {}
                         this.setTitle() != null ? <div style={this.configTitleFlagColorStyle()}
                                                        className={classes.rcsbFvRowTitleProvenanceFlag}/> : null
                     }
-                    <div className={classes.rcsbFvRowTitleText} style={{lineHeight:height+"px", paddingRight:this.PADDING_RIGHT}}>{this.setTitle()}</div>
+                    <div className={classes.rcsbFvRowTitleText+(this.state.expandTitle ? " "+classes.rcsbFvRowTitleTextExpand : "")} style={{lineHeight:height+"px", paddingRight:this.PADDING_RIGHT}} onMouseEnter={(evt)=>{this.expandTitle(evt, true)}} onMouseLeave={(evt)=>{this.expandTitle(evt, false)}}>
+                        {rowMark}
+                        {this.setTitle()}
+                    </div>
                 </div>
             );
         }
@@ -66,7 +84,7 @@ export class RcsbFvRowTitle extends React.Component <RcsbFvRowTitleInterface, {}
     /**
      * @return Title string defined in the track configuration object
      * */
-    setTitle(): string | null | ReactElement {
+    setTitle(): string | null | JSX.Element {
         if(typeof this.configData.rowTitle === "string"){
             return this.configData.rowTitle;
         }else if(typeof this.configData.rowTitle === "object"){
@@ -103,10 +121,25 @@ export class RcsbFvRowTitle extends React.Component <RcsbFvRowTitleInterface, {}
         }
         return {
             backgroundColor: color,
-            borderColor: color,
             height: this.props.rowTitleHeight,
             width: this.PADDING_RIGHT,
             float:"right"
         };
     }
+
+    private marker(): JSX.Element {
+     return (<div className={classes.rcsbFvMarker} style={{position: "relative", top: this.props.rowTitleHeight/2}} />);
+    }
+
+    private expandTitle(evt: MouseEvent<HTMLDivElement>, flag: boolean): void{
+        const div: HTMLDivElement = evt.currentTarget;
+        if ( (0 > div.clientWidth - div.scrollWidth) && flag) {
+            if(flag)
+                this.setState({expandTitle: true});
+        }else{
+            this.setState({expandTitle: false});
+        }
+
+    }
 }
+